* thread.c (ruby_thread_getcwd): fix for non-openat platforms.
* util.c (ruby_sys_getcwd): renamed and reverted.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/mvm@18209 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ iseq.c
@@ -846,7 +846,7 @@ rb_iseq_disasm_insn(VALUE ret, VALUE *iseq, size_t pos,
     }
     else {
 	rb_str_catf(str, "%04"PRIdSIZE" %-16.*s ", pos,
-		    (int)strcspn(insn_name_buff, "_"), insn_name_buff);
+		    strcspn(insn_name_buff, "_"), insn_name_buff);
     }
 
     for (j = 0; types[j]; j++) {
